refactor: 멀티 모듈 적용 누락된 파일 추가
