+++ module-menu/src/test/java/kitchenpos/menuGroup/application/MenuGroupServiceTest.java
@@ -12,7 +12,7 @@ import org.springframework.test.context.ActiveProfiles;
 import java.util.ArrayList;
 import java.util.List;
 
-import static kitchenpos.menuGroup.MenuGroupGenerator.메뉴_그룹_생성_요청;
+import static kitchenpos.generator.MenuGroupGenerator.메뉴_그룹_생성_요청;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
