+++ module-order/src/main/java/kitchenpos/order/mapper/OrderMapper.java
@@ -1,9 +1,9 @@
 package kitchenpos.order.mapper;
 
+import kitchenpos.domain.Quantity;
 import kitchenpos.menu.dao.MenuRepository;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.Menus;
-import kitchenpos.menu.domain.Quantity;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderLineItems;
